+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTitlePane.java
@@ -32,6 +32,7 @@ import java.awt.Image;
 import java.awt.Insets;
 import java.awt.Point;
 import java.awt.Rectangle;
+import java.awt.Toolkit;
 import java.awt.Window;
 import java.awt.event.ActionListener;
 import java.awt.event.ComponentEvent;
@@ -357,6 +358,7 @@ public class FlatTitlePane
 			restoreButton.setVisible( resizable && maximized );
 
 			if( maximized &&
+				!(SystemInfo.isLinux && FlatNativeLinuxLibrary.isWMUtilsSupported( window )) &&
 				rootPane.getClientProperty( "_flatlaf.maximizedBoundsUpToDate" ) == null )
 			{
 				rootPane.putClientProperty( "_flatlaf.maximizedBoundsUpToDate", null );
@@ -735,6 +737,17 @@ debug*/
 				? (state & ~Frame.ICONIFIED)
 				: (state & ~Frame.MAXIMIZED_BOTH) );
 		}
+	}
+
+	private void maximizeOrRestore() {
+		if( !(window instanceof Frame) || !((Frame)window).isResizable() )
+			return;
+
+		Frame frame = (Frame) window;
+		if( (frame.getExtendedState() & Frame.MAXIMIZED_BOTH) != 0 )
+			restore();
+		else
+			maximize();
 	}
 
 	/**
@@ -1154,23 +1167,30 @@ debug*/
 		//---- interface MouseListener ----
 
 		private Point dragOffset;
+		private boolean nativeMove;
+		private long lastSingleClickWhen;
 
 		@Override
 		public void mouseClicked( MouseEvent e ) {
+			// on Linux, when using native library, the mouse clicked event
+			// is usually not sent and maximize/restore is done in mouse pressed event
+			// this check is here for the case that a mouse clicked event comes thru for some reason
+			if( SystemInfo.isLinux && FlatNativeLinuxLibrary.isWMUtilsSupported( window ) ) {
+				// see comment in mousePressed()
+				if( lastSingleClickWhen != 0 && (e.getWhen() - lastSingleClickWhen) <= getMultiClickInterval() ) {
+					lastSingleClickWhen = 0;
+					maximizeOrRestore();
+				}
+				return;
+			}
+
 			if( e.getClickCount() == 2 && SwingUtilities.isLeftMouseButton( e ) ) {
 				if( e.getSource() == iconLabel ) {
 					// double-click on icon closes window
 					close();
-				} else if( !hasNativeCustomDecoration() &&
-					window instanceof Frame &&
-					((Frame)window).isResizable() )
-				{
+				} else if( !hasNativeCustomDecoration() ) {
 					// maximize/restore on double-click
-					Frame frame = (Frame) window;
-					if( (frame.getExtendedState() & Frame.MAXIMIZED_BOTH) != 0 )
-						restore();
-					else
-						maximize();
+					maximizeOrRestore();
 				}
 			}
 		}
@@ -1180,10 +1200,56 @@ debug*/
 			if( window == null )
 				return; // should newer occur
 
+			// on Linux, show window menu
+			if( SwingUtilities.isRightMouseButton( e ) &&
+				SystemInfo.isLinux && FlatNativeLinuxLibrary.isWMUtilsSupported( window ) )
+			{
+				e.consume();
+				FlatNativeLinuxLibrary.showWindowMenu( window, e );
+				return;
+			}
+
 			if( !SwingUtilities.isLeftMouseButton( e ) )
 				return;
 
 			dragOffset = SwingUtilities.convertPoint( FlatTitlePane.this, e.getPoint(), window );
+			nativeMove = false;
+
+			// on Linux, move or maximize/restore window
+			if( SystemInfo.isLinux && FlatNativeLinuxLibrary.isWMUtilsSupported( window ) ) {
+				// The fired Java mouse events, when doing a double-click and the first click
+				// sends a _NET_WM_MOVERESIZE message, are different for various Linux distributions:
+				//   CentOS 7      (GNOME 3.28.2, X11):   PRESSED(clickCount=1) PRESSED(clickCount=2) RELEASED(clickCount=2)
+				//   Ubuntu 20.04  (GNOME 3.36.1, X11):   PRESSED(clickCount=1) PRESSED(clickCount=2) RELEASED(clickCount=2)
+				//   Ubuntu 22.04  (GNOME 42.2, Wayland): PRESSED(clickCount=1)                       RELEASED(clickCount=1) CLICKED(clickCount=1)
+				//   Kubuntu 22.04 (KDE 5.24.4, X11):     PRESSED(clickCount=1) PRESSED(clickCount=1) RELEASED(clickCount=1)
+
+				// double-click is not always recognized in Java when using _NET_WM_MOVERESIZE message
+				int clickCount = e.getClickCount();
+				if( clickCount == 1 && lastSingleClickWhen != 0 && (e.getWhen() - lastSingleClickWhen) <= getMultiClickInterval() )
+					clickCount = 2;
+
+				switch( clickCount ) {
+					case 1:
+						// move window via _NET_WM_MOVERESIZE message
+						e.consume();
+						nativeMove = FlatNativeLinuxLibrary.moveOrResizeWindow( window, e, FlatNativeLinuxLibrary.MOVE );
+						lastSingleClickWhen = e.getWhen();
+						break;
+
+					case 2:
+						// maximize/restore on double-click
+						// also done here because no mouse clicked event is sent when using _NET_WM_MOVERESIZE message
+						lastSingleClickWhen = 0;
+						maximizeOrRestore();
+						break;
+				}
+			}
+		}
+
+		private int getMultiClickInterval() {
+			Object value = Toolkit.getDefaultToolkit().getDesktopProperty( "awt.multiClickInterval" );
+			return (value instanceof Integer) ? (Integer) value : 500;
 		}
 
 		@Override public void mouseReleased( MouseEvent e ) {}
@@ -1194,8 +1260,11 @@ debug*/
 
 		@Override
 		public void mouseDragged( MouseEvent e ) {
-			if( window == null )
+			if( window == null || dragOffset == null )
 				return; // should newer occur
+
+			if( nativeMove )
+				return;
 
 			if( !SwingUtilities.isLeftMouseButton( e ) )
 				return;